refactor: Move TripAlteration to new transit model
