+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/EnumTypes.java
@@ -4,7 +4,6 @@ import graphql.schema.GraphQLEnumType;
 import java.util.Arrays;
 import java.util.function.Function;
 import org.opentripplanner.model.BookingMethod;
-import org.opentripplanner.model.TripAlteration;
 import org.opentripplanner.model.plan.AbsoluteDirection;
 import org.opentripplanner.model.plan.RelativeDirection;
 import org.opentripplanner.model.plan.VertexType;
@@ -22,6 +21,7 @@ import org.opentripplanner.routing.trippattern.RealTimeState;
 import org.opentripplanner.transit.model.network.BikeAccess;
 import org.opentripplanner.transit.model.network.TransitMode;
 import org.opentripplanner.transit.model.timetable.Direction;
+import org.opentripplanner.transit.model.timetable.TripAlteration;
 
 public class EnumTypes {
