+++ src/main/java/org/opentripplanner/model/TripOnServiceDate.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.framework.TransitEntity;
 import org.opentripplanner.transit.model.timetable.Trip;
+import org.opentripplanner.transit.model.timetable.TripAlteration;
 
 /**
  * Class for holding data about a certain trip on a certain day. Essentially a DatedServiceJourney.
